컴포넌트 분리 및 console.log제거
